v2.5.1 patch
The following was fixed or added in SMP patch 2.5.1:
Decoding and processing of the tag type 7 data (free form field) when
checking a security label in the SecurityLabel::check method in ACL.
Bug fix in tracking of required categories in the
SecurityLabel::requiredCatCheck method in ACL.
Bug fix in local ACL cache when storing and retrieving attribute
certificates as cacheable objects.
Bug fix in revocation hash table deletion code in revocation status
checking library.