v2.5.2 PATCH (October 3, 2006)
The following was fixed or added in SMP patch 2.5.2:
Corrected a bug where certificates added to the database from a URL are
not found locally with the SRL_RequestObjs() function.
Corrected a bug where certificates added to the database from a URL do
not include the flag to include the crossCertificatePair directory
attribute.
Set the default LDAP client behavior from v2 to v3.
Removed memory leaks from the ACL and SFL libraries.
Fixed the incorrect algorithm in Findcrlissuer() function that was
causing performance problems.
Fixed a code bug where the base class for a CRL is added to the cache
instead of the derived class.
Delta CRL processing now occurs if the freshest extension is in the
CRL, not just in the certificate.
CRL processing code now works around not having an AsnInt::operator >()
function by using < and == operators. Also fixed a problem with a delta
CRL being issued at the same time as a base CRL.
Added support to more extended key usages in the CML.
Removed a memory leak from the ACL.
The code now considers a delta CRL applicable to a base CRL if the
CRLNumbers in the delta and base are equal.
Fixed a problem with CRL updates to the database using improper hash
values.
Can now build CML into static libraries.
Various fixes in the ACL, CML, and SFL libraries.
Now supports Crypto++ 5.2.1 and GCC v4.1.1 and below.
Fixed a bug in the dbu_read_entry.
The patch has been tested on Linux/Windows without issue.